UserProfileComponent 에서 get me api를 호출하지 않고, MainComponent에서 userInfo를 얻은 후 각 Component로 뿌려주는 방식으로 구조를 개편했습니다.
